+++ NEWS.md
@@ -1,4 +1,13 @@
 # log history of echarty package development
 
-## echarty v0.1.0
-* initial release Feb 2021
+## echarty 0.1.1
+
+- added ec.plugjs - install unknown plugins, like JS maps
+- added ec.js2r - a JS to R translation assistant, a Shiny app
+- ec.init update: one-time install for all plugins except built-in leaflet/custom
+- GL merged with 3D as plugin '3D'
+- more plugins: liquidfill, world, wordcloud, gmodularity
+- ec.sband renamed to ec.band 
+
+## echarty 0.1.0
+- initial release, approved by CRAN Feb 16, 2021